Switch back to 32-bit operation.
Switch to 64-bit operation in 4c3389b56f08a3c1532a915ace6ca429a7051b97
was my mistake.

+++ ir_fold.h
@@ -394,7 +394,7 @@ IR_FOLD(SUB(C_I16, C_I16))
 IR_FOLD(SUB(C_I32, C_I32))
 {
 	IR_ASSERT(IR_OPT_TYPE(opt) == op1_insn->type);
-	IR_FOLD_CONST_I((int32_t)(op1_insn->val.u64 - op2_insn->val.u64));
+	IR_FOLD_CONST_I((int32_t)(op1_insn->val.u32 - op2_insn->val.u32));
 }
 
 IR_FOLD(SUB(C_I64, C_I64))
